requirements.txt
torch
torchvision
transformers
kafka-python
easyocr
pillow
numpy
opencv-python
requests
fastapi
uvicorn
websockets
